updated create-ticket
